+++ packages/web3-core/src/types.ts
@@ -15,7 +15,6 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { HexString } from 'web3-types';
+import { HexString, Transaction } from 'web3-types';
 
-// TODO: When we have `web3-types` package we can share TransactionType
-export type TransactionTypeParser = (transaction: Record<string, unknown>) => HexString | undefined;
+export type TransactionTypeParser = (transaction: Transaction) => HexString | undefined;