+++ packages/web3-eth-ens/test/integration/ens.events.test.ts
@@ -17,10 +17,10 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Contract, PayableTxOptions } from 'web3-eth-contract';
-import { sha3, DEFAULT_RETURN_FORMAT } from 'web3-utils';
+import { sha3 } from 'web3-utils';
 import { getBlock } from 'web3-eth';
 
-import { Address, Bytes } from 'web3-types';
+import { Address, Bytes, DEFAULT_RETURN_FORMAT } from 'web3-types';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import IpcProvider from 'web3-providers-ipc';
 import { ENS } from '../../src';