+++ packages/web3-eth/test/integration/block/rpc.getBlock.test.ts
@@ -14,8 +14,13 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { FMT_BYTES, FMT_NUMBER } from 'web3-utils';
-import { TransactionReceipt, Transaction, SupportedProviders } from 'web3-types';
+import {
+	TransactionReceipt,
+	Transaction,
+	FMT_BYTES,
+	FMT_NUMBER,
+	SupportedProviders,
+} from 'web3-types';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Contract } from 'web3-eth-contract';
 import { validator } from 'web3-validator';