+++ packages/web3-eth/test/integration/defaults.test.ts
@@ -16,7 +16,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Contract } from 'web3-eth-contract';
-import { hexToNumber, numberToHex, DEFAULT_RETURN_FORMAT } from 'web3-utils';
+import { hexToNumber, numberToHex } from 'web3-utils';
 import { TransactionBuilder, TransactionTypeParser, Web3Context, Web3PromiEvent } from 'web3-core';
 import {
 	Hardfork,
@@ -24,6 +24,7 @@ import {
 	TransactionReceipt,
 	ValidChains,
 	Web3BaseProvider,
+	DEFAULT_RETURN_FORMAT,
 } from 'web3-types';
 import {
 	detectTransactionType,