+++ packages/web3-eth/test/integration/rpc.test.ts
@@ -20,17 +20,12 @@ import {
 	TransactionReceipt,
 	TransactionInfo,
 	SupportedProviders,
+	FMT_BYTES,
+	FMT_NUMBER,
 } from 'web3-types';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Contract, decodeEventABI } from 'web3-eth-contract';
-import {
-	hexToNumber,
-	hexToString,
-	numberToHex,
-	FMT_BYTES,
-	FMT_NUMBER,
-	getStorageSlotNumForLongString,
-} from 'web3-utils';
+import { hexToNumber, hexToString, numberToHex, getStorageSlotNumForLongString } from 'web3-utils';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Web3Eth } from '../../src';