+++ packages/web3-types/src/eth_types.ts
@@ -220,6 +220,7 @@ export interface SyncOutput {
 export type Receipt = Record<string, unknown>;
 
 type FilterOption = Record<string, Numbers | Numbers[]>;
+
 // https://github.com/ethereum/execution-apis/blob/main/src/schemas/filter.json#L28
 export interface Filter {
 	readonly fromBlock?: BlockNumberOrTag;